Minor optimizations
- Remove ancient flag in transposition table, since now it's reset in every search
- Change Centipawns type from isize to i32 (save 64 MB of memory in transposition table)
- Change history_moves from u16 to Centipawns (use extra 128 kB of memory, avoid overflows)
- Replace king score with the one it would have if it was a custom piece (in preparation for removing all non-custom pieces). The checkmate score is still -99999 centipawns

+++ protochess-engine-rs/src/types/piece_type.rs
@@ -1,6 +1,6 @@
 
 // Class for storing the value of a position, must allow negative values
-pub type Centipawns = isize;
+pub type Centipawns = i32;
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug, Copy)]
 pub enum PieceType {
